% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_default_samples}
\alias{add_default_samples}
\alias{add_default_samples.numeric}
\alias{add_default_samples.data.frame}
\alias{add_default_samples.tbl_spark}
\alias{add_default_samples.modeler}
\title{Add Default Samples function}
\usage{
add_default_samples(x, ...)

\method{add_default_samples}{numeric}(x)

\method{add_default_samples}{data.frame}(df)

\method{add_default_samples}{tbl_spark}(df)

\method{add_default_samples}{modeler}(obj)
}
\arguments{
\item{x}{numeric vector}

\item{df}{R data.frame}

\item{obj}{modeler object}

\item{df}{Spark Dataframe}
}
\description{
Function creates a no-frills sampling object. Object has a single train index
only. No validation or test indicies created.
}
\details{
Function used when modeler object create for place holder for modeler object.
Default sampling can be used for simple out of the box models
}
