% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_holdout_samples}
\alias{add_holdout_samples}
\alias{add_holdout_samples.integer}
\alias{add_holdout_samples.data.frame}
\alias{add_holdout_samples.tbl_spark}
\alias{add_holdout_samples.modeler}
\title{Add Holdout Samples function}
\usage{
add_holdout_samples(x, splits)

\method{add_holdout_samples}{integer}(x, splits)

\method{add_holdout_samples}{data.frame}(x, splits)

\method{add_holdout_samples}{tbl_spark}(x, splits)

\method{add_holdout_samples}{modeler}(x, splits)
}
\arguments{
\item{x}{numeric vector, data.frame, spark dataframe, or modeler object to
create holdout samples from}

\item{splits}{numeric vector of holdout splits. Values need to sum to 1. Each
value represents the amount of data partitioned. Order matters - the splits
correspond to train/validation/test*. Test value optional.}
}
\value{
Returns updated modeler object if modeler object provided otherwise
returns samples object
}
\description{
Function creates a new samples object with holdout samples based on
configuration
}
\details{
Holdout samples take sequential row indicies. Holdout samples are used
primarily in forecasting applications but can be used for other modeling
applications
}
\examples{

# Data.frame example
add_holdout_samples(mtcars, splits = c(.8, .2))
}
