% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_model}
\alias{add_model}
\title{Add Model to Modeler Object function}
\usage{
add_model(obj, pipeline = pipeline(), method, ..., desc = NULL,
  path = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{pipeline}{pipeline object. default is empty pipeline which applies no
data transformations}

\item{method}{string input of model method}

\item{...}{additional arguments to pass to model method}

\item{desc}{optional model description}

\item{path}{optional file path to save model}
}
\value{
modeler object with model added
}
\description{
Function to add model to modeler object. More than one model can be added to
a modeler object.
}
\details{
Function creates a new model object from inputs and then appends to modeler
models list
}
