% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_models}
\alias{add_models}
\title{Add Multiple Models to Modeler Object function}
\usage{
add_models(obj, pipe = NULL, models, class)
}
\arguments{
\item{obj}{modeler object}

\item{pipe}{pipeline object}

\item{models}{list with models method and list of arguments in each element}

\item{class}{modeler object class}
}
\value{
modeler object with models added
}
\description{
Function to add model to modeler object. More than one model can be added to
a modeler object.
}
\details{
Function creates a new model object from inputs and then appends to modeler
models list
}
