% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_resample_samples}
\alias{add_resample_samples}
\alias{add_resample_samples.numeric}
\alias{add_resample_samples.data.frame}
\alias{add_resample_samples.tbl_spark}
\alias{add_resample_samples.modeler}
\alias{add_resample_samples.forecaster}
\title{Add Random Resamples function}
\usage{
add_resample_samples(x, number, amount, test_holdout_prct, seed)

\method{add_resample_samples}{numeric}(x, number, amount,
  test_holdout_prct = NULL, seed = NULL)

\method{add_resample_samples}{data.frame}(x, number, amount,
  test_holdout_prct = NULL, seed = NULL)

\method{add_resample_samples}{tbl_spark}(x, number, amount,
  test_holdout_prct = NULL, seed = NULL)

\method{add_resample_samples}{modeler}(x, number, amount,
  test_holdout_prct = NULL, seed = NULL)

\method{add_resample_samples}{forecaster}(x, number, amount,
  test_holdout_prct = NULL, seed = NULL)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create samples from}

\item{number}{number of random resamples}

\item{amount}{amount of data used for train sample. 1-amount used for
validation sample.}

\item{test_holdout_prct}{amount of data to holdout out for test sample}

\item{seed}{random seed generator}
}
\value{
either an updated modeler object if provided otherwise a samples
  object with resamples
}
\description{
Function creates a new samples object with random resamples based on
configuration. Each index is random sample of the dataset
}
\details{
The function creates train and validation pairs using all of the dataset rows
}
\examples{

# Data.frame example
add_resample_samples(mtcars, number = 5, amount = .5)
}
