% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{append_model}
\alias{append_model}
\title{Append Model to Modeler Object function}
\usage{
append_model(obj, model)
}
\arguments{
\item{obj}{modeler object}

\item{model}{model object}
}
\value{
modeler object with model added
}
\description{
Function to append a valid model to a modeler object.
}
\details{
Function updates the model status and appends to modeler models list
}
