% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{auto_forecaster}
\alias{auto_forecaster}
\title{Auto Forecaster Function}
\usage{
auto_forecaster(df, target, index_var, periods, unit = NULL, pipe = NULL,
  models = list(list(method = "auto.arima", method_args = list()), list(method
  = "ets", method_args = list())), splits = c(0.8, 0.2), conf_levels = c(80,
  95))
}
\arguments{
\item{df}{dataframe}

\item{target}{numeric variable to model and forecast}

\item{index_var}{index variable}

\item{periods}{number of periods to forecast}

\item{unit}{unit of index variable default is null}

\item{pipe}{pipeline object}

\item{models}{nested list of models. each model list should have a method and
list of arguments}

\item{splits}{holdout splits ratios default is 80-20 train to validation}

\item{conf_levels}{forecast confidence levels. default is 80 and 90 percent}
}
\value{
predictions object
}
\description{
Creates automated forecasts for univariate time series.
}
\details{
Convience wrapper to create complete forecaster pipeline
}
