% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{auto_fourier}
\alias{auto_fourier}
\title{Fit Auto Fouier Arima model to univariate time series}
\usage{
auto_fourier(y, xreg = NULL, frequency = NULL, ic = "aicc", ...)
}
\arguments{
\item{y}{a univariate time series}

\item{xreg}{Optionally, a vector or matrix of external regressors, which
must have the same number of rows as \code{y}.}

\item{frequency}{numeric input for seasonal frequency of y - univarite time
series. Default is NULL for no seasonality}

\item{ic}{Fit metric to select best model. Options are 'aic', 'aicc',
'loglik', or 'bic'. Default is 'aicc'}

\item{...}{Additional arguments to be passed to \code{\link[stats]{arima}}.}

\item{x}{Optional a vector or matrix of extrenal regressors. Must have same
number of rows as y}
}
\value{
returns a list with best ARIMA model according AIC including harmonic
  regressors and K number of harmonic regressor terms
}
\description{
The auto_fourier_arima function is a wrapper function around the auto.arima
function in the forecast package. The wrapper adds fourier harmonic
regressors to model seasonal factors. This function fits multiple fourier
frequencies and selects the model with the best fit as measured by min aic.
The case of no harmonic regressors is also considered
}
\examples{
library(forecast)
auto_fourier(USAccDeaths, frequency = 12)
}
