% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlater.R
\name{corr}
\alias{corr}
\alias{corr.data.frame}
\alias{corr.tbl_spark}
\title{DataFrame Correlation Function}
\usage{
corr(df, ...)

\method{corr}{data.frame}(df, target_var = NULL,
  target_var_name = "target_variable")

\method{corr}{tbl_spark}(df, target_var = NULL,
  target_var_name = "target_variable")
}
\arguments{
\item{df}{DataFrame}

\item{target_var}{optional input to compute correlation for a single target
column producing a 1xN vector of values. Default value is NULL which
produces a NxN matrix of values.}

\item{target_var_name}{optional column name for id column. default is
'target_variable'. this differs from stats::cor function which returns NxN
matrix with column and row names. First column returned by corr is
equivalent to rownames in cor function}
}
\value{
Dataframe of either 1xN vector or NxN martrix with pearson
 correlation values with additional rowname id column
}
\description{
Function that calculates the pearson correlation for one or more pairs of
numerical data types.
}
\details{
This function is used in the correlater function but can be used as stand
alone function.
}
\examples{
library(dplyr)
corr(mtcars)
corr(mtcars, "mpg")
corr(mtcars, "mpg", "id_var")
}
