% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlater.R
\name{correlater}
\alias{correlater}
\alias{correlater.data.frame}
\alias{correlater.tbl_spark}
\title{DataFrame Correlater function}
\usage{
correlater(df, ...)

\method{correlater}{data.frame}(df, target_var = NULL, transform = NULL,
  output_col_names = c("var1", "var2", "cor"), remove_diag = FALSE)

\method{correlater}{tbl_spark}(df, target_var = NULL, transform = NULL,
  output_col_names = c("var1", "var2", "cor"), remove_diag = FALSE,
  collect = TRUE)
}
\arguments{
\item{df}{DataFrame}

\item{target_var}{optional input to compute correlation for a single target
column producing a 1xN vector of values. Default value is NULL which
produces a NxN matrix of values.}

\item{transform}{optional input to transform data prior to correlation
calculation. Default is NULL (no transformation). Options are  'standardize'
or 'normalize'}

\item{output_col_names}{colnames of output dataset. default is c("var1",
"var2", "cor")}

\item{remove_diag}{optional binary T/F input for removing diagonal values from
the correlation matrix}

\item{collect}{logical option to collect transformed dataset into R prior to
corr function. applies to Spark DataFrames only}
}
\value{
Dataframe with pairwise correlation values in narrow, long format
}
\description{
Function that calculates the pearson correlation for a dataset. Function
removes non-numerical columns, calculates pairwise pearson correlation values
for 1 or more columns, and transforms the results into long format.
}
\details{
The function provides an option to transform the data prior to the correlation
calculation with either the standardizer or normalizer transformation
functions. The target_variable input allows the user to compute correlation
values for a single target variable, which can reduce unwanted or
un-neccessary computations.
}
