% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexer.R
\name{date_parts}
\alias{date_parts}
\title{Date Parts}
\usage{
date_parts(x, label = FALSE, abbr = FALSE)
}
\arguments{
\item{x}{Date values}

\item{label}{logical option to change month and day of week values to label.
default is FALSE}

\item{abbr}{logical option to use abbreviate labels. default is FALSE}
}
\description{
Function to calculate date parts for Date
}
