% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter.R
\name{detect}
\alias{detect}
\alias{detect.data.frame}
\alias{detect.grouped_df}
\title{Dataframe Anomaly Detection function}
\usage{
detect(df, ...)

\method{detect}{data.frame}(df, index_var = NULL, measure_var = NULL,
  frequency, direction = "pos", alpha = 0.01, max_anoms = 0.01,
  trend_window = 0.75)

\method{detect}{grouped_df}(df, index_var, measure_var, frequency,
  direction = "pos", alpha = 0.01, max_anoms = 0.01,
  trend_window = 0.75)
}
\arguments{
\item{df}{dataframe}

\item{index_var}{index date/datetime variable. Default is NULL which will set
the index variable to the first column in the dataframe}

\item{measure_var}{numeric measure variable to analyze. Default is NULL which
will set the measure variable to the second column in the dataframe}

\item{frequency}{numeric period length of seasonality. Ex - 7 for daily data
with weekly seasonality}

\item{direction}{character input for direction of anomalies to detect.
Options are 'pos' - only positive anomalies are detected, 'neg' only
negative anomalies are detected, or 'both' for both 'pos' and 'neg'
anomalies}

\item{alpha}{numeric statistical threshold for anomlie detection. default is
.01 or 99 percent Confidence.}

\item{max_anoms}{numeric threshold for percent of data identified as
anomalies. default is .01 or 1 percent of the data}

\item{trend_window}{numeric input for percentage of total data width for
trend detection. default is 0.75. Large values smooth the trend which may
not be appropriate for non-stationary trends. Small values may overfit the
trend and lead to may false negatives}
}
\value{
dataframe with same number of records as input dataframe. Appends
  additional fields for seasonality, trend, residual, and anomaly. Anomalies
  are coded with a binary flag of 1
}
\description{
Anomaly detection function for time series data. Function determines
statistical anomlies in the dataset based on configurable theshold
parameters.
}
\details{
Function is similiar to Seasonl Hybrid ESD algorithm discussed
here \url{https://arxiv.org/pdf/1704.07706.pdf} and implemented in the
\code{\link[AnomalyDetection]{AnomalyDetectionTs}}
}
