% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter.R
\name{detecter}
\alias{detecter}
\alias{detecter.data.frame}
\alias{detecter.tbl_spark}
\title{Anomaly Dectection Wrapper Function}
\usage{
detecter(df, ...)

\method{detecter}{data.frame}(df, index_var, group_vars = NULL, measure_vars,
  frequency, direction = "pos", alpha = 0.01, max_anoms = 0.01,
  trend_window = 0.75)

\method{detecter}{tbl_spark}(df, index_var, group_vars, measure_vars, frequency,
  direction = "pos", alpha = 0.01, max_anoms = 0.01,
  trend_window = 0.75)
}
\arguments{
\item{df}{dataframe}

\item{index_var}{index date/datetime variable. Default is NULL which will set
the index variable to the first column in the dataframe}

\item{group_vars}{optional vector of grouping variables to apply dectect
function to. default is NULL - no grouping}

\item{frequency}{numeric period length of seasonality. Ex - 7 for daily data
with weekly seasonality}

\item{direction}{character input for direction of anomalies to detect.
Options are 'pos' - only positive anomalies are detected, 'neg' only
negative anomalies are detected, or 'both' for both 'pos' and 'neg'
anomalies}

\item{alpha}{numeric statistical threshold for anomlie detection. default is
.01 or 99% Confidence.}

\item{max_anoms}{numeric threshold for percent of data identified as
anomalies. default is .01 or 1% of the data}

\item{trend_window}{numeric input for percentage of total data width for
trend detection. default is 0.75. Large values smooth the trend which may
not be appropriate for non-stationary trends. Small values may overfit the
trend and lead to may false negatives}
}
\description{
Detecter is a wrapper function for the detect function. Detecter takes a
normalized index based dataframe and reshapes for core detect function.
Detecter allows for grouping and multiple measure variables
}
