% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R, R/spark-ml.R
\name{evaluate.forecast_model}
\alias{evaluate.forecast_model}
\alias{evaluate}
\alias{evaluate.spark_ml_clustering}
\title{Evaluate Model Generic}
\usage{
\method{evaluate}{forecast_model}(mobj, target_df, measure)

evaluate(mobj, target_df, measure)

\method{evaluate}{spark_ml_clustering}(mobj, measure)
}
\arguments{
\item{mobj}{model object}

\item{target_df}{dataframe with target and index variables}

\item{measure}{measure object}
}
\value{
returns evaluted model object
}
\description{
Function to evaluate the predictive performance of a model

Evaluate Method for Spark-ML Clustering Object
}
