% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{execute}
\alias{execute}
\alias{execute.tbl_spark}
\alias{execute.modeler}
\title{Pipeline Execute function}
\usage{
execute(x, pipe)

\method{execute}{tbl_spark}(x, pipe)

\method{execute}{modeler}(x, pipe)
}
\arguments{
\item{x}{input object to apply pipeline expression on. Can either be
data.frame, spark dataframe or modeler object}

\item{pipe}{pipeline object}
}
\value{
updated pipeline object with output
}
\description{
Execute function executes the pipeline expression on the input and stores in
output location
}
\examples{

pipe <- pipeline(expr = function(e) mean(e$mpg))
execute(mtcars, pipe)
}
