% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expander.R
\name{expander}
\alias{expander}
\alias{expander.data.frame}
\alias{expander.tbl_spark}
\title{Dataframe Expander Function}
\usage{
expander(df, group_vars, id_vars, fun, mode, complete)

\method{expander}{data.frame}(df, group_vars = NULL, id_vars = NULL,
  fun = NULL, mode = "crossing", complete = TRUE)

\method{expander}{tbl_spark}(df, group_vars = NULL, id_vars = NULL,
  fun = NULL, mode = "crossing", complete = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{group_vars}{optional vector column names to group data by. default is
NULL. Any expansion function will be nested within the grouping vars}

\item{id_vars}{optional vector of colnames to expand data by. default it
NULL.}

\item{fun}{optional expansion function. useful for expanding continuous
variables with values that don't appear in the data (like missing dates).
fun argument should be wrapped in `funs(...)` function}

\item{mode}{expansion mode for id_vars arguments. `nesting`` mode returns all
unique combinations of id_vars seen in the data. `crossing` mode returns
all combinations of the unique values of id_vars. `crossing` mode similar
to `expand.grid` in base R}

\item{complete}{logical option to return original data `right_join` to
expanded dataset. FALSE returns just expanded grid. Default is TRUE}
}
\value{
dataframe
}
\description{
Function to expand dataset by converting implicit missing values to explict
values
}
\examples{
library(tidyr)
library(dplyr)
expander(mtcars, id_vars = c('am', 'cyl'), complete = FALSE)
expander(mtcars, id_vars = c('am', 'cyl'), fun = funs(carb = 1:5), complete = TRUE)
expander(mtcars, id_vars = c('am'), fun = funs(cyl = full_seq(cyl, 1)), complete = TRUE)
}
\seealso{
\url{http://tidyr.tidyverse.org/reference/expand.html} for tidyr `expand` function which is used
  internally by `expander`
}
