% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R, R/spark-ml.R
\name{fit.forecast_model}
\alias{fit.forecast_model}
\alias{fit}
\alias{fit.spark_ml}
\title{Fit Model Generic}
\usage{
\method{fit}{forecast_model}(mobj, data, ...)

fit(mobj, data, ...)

\method{fit}{spark_ml}(mobj, data)
}
\arguments{
\item{mobj}{Model object to fit}

\item{data}{data to fit model object on}

\item{...}{additional arguments to pass through}
}
\description{
Fit model to single data sample
}
