% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{fit_args.forecast}
\alias{fit_args.forecast}
\title{Forecast Fit Arguments Helper function}
\usage{
\method{fit_args}{forecast}(model = "auto.arima", df, index_var, y_var,
  x_vars = NULL, boxcox = FALSE, ...)
}
\description{
Forecast Fit Arguments Helper function
}
