% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{fit_model.forecaster_fit_args}
\alias{fit_model.forecaster_fit_args}
\title{Forecast Fit Model function}
\usage{
\method{fit_model}{forecaster_fit_args}(obj)
}
\arguments{
\item{obj}{forecast fit argument object. created with fit_args method.}
}
\value{
forecaster model object
}
\description{
Forecast Fit Model function
}
