% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{flow}
\alias{flow}
\alias{flow.tbl_spark}
\alias{flow.modeler}
\title{Pipeline Flow function}
\usage{
flow(x, pipe)

\method{flow}{tbl_spark}(x, pipe)

\method{flow}{modeler}(x, pipe)
}
\arguments{
\item{x}{input object to apply pipeline expression on. Can either be
data.frame, spark dataframe or modeler object}

\item{pipe}{pipeline object}
}
\value{
updated pipeline object with output
}
\description{
Flow function executes the pipeline expression on the input and stores in
output location
}
\examples{

pipe <- pipeline(expr = function(e) mean(e$mpg))
flow(mtcars, pipe)
}
