% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{flush}
\alias{flush}
\title{Pipeline Flush function}
\usage{
flush(pipe)
}
\arguments{
\item{pipe}{pipeline object}
}
\value{
updated pipeline object with data output removed
}
\description{
Flush function removes any stored output data from a pipeline
}
\details{
Function useful for cleaning up and reducing size of modeler object
}
\examples{

pipe <- pipeline(expr = function(e) mean(e$mpg))
pipe <- execute(mtcars, pipe)
clean_pipe <- flush(pipe)
clean_pipe$output
}
