% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{flush_pipes}
\alias{flush_pipes}
\title{Flush Pipelines function}
\usage{
flush_pipes(obj, ids = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{ids}{one or more model ids. default is null which flushes all pipes}
}
\value{
updated modeler object
}
\description{
Function to flush pipelines for one or more models in modeler object
}
