% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{forecast.auto_fourier}
\alias{forecast.auto_fourier}
\title{Forecast Auto Fourier Arima model for univariate time series}
\usage{
forecast.auto_fourier(object, h = 10, xreg = NULL, frequency,
  level = c(80, 95), ...)
}
\arguments{
\item{xreg}{Optionally, a vector or matrix of external regressors, which
must have the same number of rows as \code{y}.}

\item{frequency}{seasonal frequency of y - univarite time series}

\item{...}{Additional arguments to be passed to \code{\link[stats]{arima}}.}

\item{x}{Optional a vector or matrix of extrenal regressors. Must correspond
to forecast period and be the same length as the forecast}
}
\value{
An object of class 'forecast'
}
\description{
Function to provide forecasts from an auto_fourier_arima model. The
auto_fourier_model adds harmonic regressors to the auto_arima function from
the forecast function
}
\examples{
library(forecast)
fit <- auto_fourier(USAccDeaths, 12)
auto_fourier_f(fit$model, frequency=12, h=12)
}
