% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{forecaster}
\alias{forecaster}
\title{Forecaster Constructer Function}
\usage{
forecaster(df, target, name = NULL, id = NULL, version = NULL,
  desc = NULL, scientist = NULL, dir = NULL, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{target}{target variable for supervised use cases such as forecaster,
regresser, and classifier}

\item{name}{optional name for use case. Default is to append target with type}

\item{id}{optional id string. A random id will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{dir}{optional directory path for saving reports, data and models}

\item{...}{any addtional parameters}
}
\description{
Creates a forecaster object which inherits from modeler class
}
\seealso{
Other use cases: \code{\link{modeler}}
}
\concept{use cases}
