% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{get_coefs.forecast_model}
\alias{get_coefs.forecast_model}
\alias{get_coefs}
\title{Get Coefficients from Forecast Model Object}
\usage{
\method{get_coefs}{forecast_model}(mobj)

get_coefs(x, ...)
}
\arguments{
\item{mobj}{forecast model object as a result of fit function}
}
\description{
Get Coefficients from Forecast Model Object

Return the Model Coefficients
}
