% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{get_fit.forecast_model}
\alias{get_fit.forecast_model}
\alias{get_fit.forecast_predict}
\title{Get Fitted Values from Forecast Model Object}
\usage{
\method{get_fit}{forecast_model}(mobj)

\method{get_fit}{forecast_predict}(pobj)
}
\arguments{
\item{mobj}{forecast model object as a result of fit function}

\item{pobj}{forecast prediction object as a result of call to predict function}
}
\value{
data.frame with forecast model fitted values
}
\description{
Get Fitted Values from Forecast Model Object

Function to Get Fit from forecast prediction
}
