% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{get_forecasts.forecast}
\alias{get_forecasts.forecast}
\alias{get_forecasts}
\title{Get Forecasts from forecast object}
\usage{
\method{get_forecasts}{forecast}(fobj)

get_forecasts(x, ...)
}
\arguments{
\item{fobj}{forecast object}
}
\value{
data.frame with forecats from forecast model
}
\description{
Get Forecasts from forecast object

Get Model Forecasts
}
