% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{get_geom_params}
\alias{get_geom_params}
\title{Get GGPlot2 geom Parameters}
\usage{
get_geom_params(names, fill = NULL, color = NULL, linetype = NULL,
  size = NULL, alpha = NULL, group = NULL, shape = NULL,
  aes_params = list(), geom_params = list())
}
\arguments{
\item{names}{colnames of the data set being used for charting. expects a
vector of string values}

\item{fill}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to fill the histogram
with color. expects a string.}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string.}

\item{linetype}{Line types should be specified with either an integer, a name, or
with a string of an even number (up to eight) of hexadecimal digits which give
the lengths in consecutive positions in the string.
0 = blank, 1 = solid, 2 = dashed, 3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash}

\item{size}{Should be specified with a numerical value (in millimetres),
or from a variable source}

\item{alpha}{sets the opacity of the points}

\item{group}{a variable name to apply a group by operation and create
subgroups of line charts}

\item{shape}{Shape takes four types of values: an integer in [0, 25],
a single character-- which uses that character as the plotting symbol,
a . to draw the smallest rectangle that is visible (i.e., about one pixel)
an NA to draw nothing}

\item{aes_params}{list of parameters to pass to the aes function.
default is empty list}

\item{geom_params}{list of parameters to pass to the geom function. default
is empty list}
}
\value{
returns a list of 2 lists. First returned list is the updated
  arguments to pass to the aes function. the second list is the updated list
  of arguments to pass to the geom function
}
\description{
Function evaluates, parses, and sets the arguments for the
\code{\link{ggplot2::aes}} function and the geom function.
}
\details{
Function required to handle fill and color arguments that can be both a
variable name or a color value

Internal function for the sncr_* family of charting functions
}
