% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_indicies}
\alias{get_indicies}
\alias{get_indicies.samples}
\alias{get_indicies.modeler}
\title{Get Sample Indicies}
\usage{
get_indicies(obj, ...)

\method{get_indicies}{samples}(obj)

\method{get_indicies}{modeler}(obj)
}
\arguments{
\item{...}{}
}
\value{
list with train and validation indicies pairs
}
\description{
Function to extract the train and validation indicies pairs
}
\details{
Returns a list with an element for each train and validation indicies pairs.
Each pair used in model evaluation step where a model is fit to the train
index and evaluated on the predictions made on the validation index
}
\examples{
# Simple example
library(dplyr)
mtcars \%>\%
add_holdout_samples(., splits = c(.6, .2, .2)) \%>\%
get_indicies(.)
}
