% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/spark-ml.R
\name{tidy_performance.forecast_model}
\alias{tidy_performance.forecast_model}
\alias{tidy_performance.spark_ml}
\title{Tidy Forecast Model Performance Object}
\usage{
\method{tidy_performance}{forecast_model}(mobj)

\method{tidy_performance}{spark_ml}(mobj)
}
\arguments{
\item{mobj}{forecast model object as a result of fit function}

\item{mobj}{forecast model object as a result of fit function}
}
\description{
Tidy Forecast Model Performance Object

Tidy Forecast Model Performance Object
}
