% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_predictions.forecast_predict}
\alias{get_predictions.forecast_predict}
\title{Function to Get Predictions from forecast prediction}
\usage{
\method{get_predictions}{forecast_predict}(pobj)
}
\arguments{
\item{pobj}{forecast prediction object as a result of call to predict function}
}
\value{
data.frame with forecast predictions
}
\description{
Function to Get Predictions from forecast prediction
}
