% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/modeler.R
\name{get_target.model}
\alias{get_target.model}
\alias{get_target}
\title{Get Target Data function}
\usage{
get_target.model(obj)

get_target(obj)
}
\arguments{
\item{obj}{modeler object}
}
\description{
Returns modeler target data. Calls get_target_df generic
}
