% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_test_samples}
\alias{get_test_samples}
\alias{get_test_samples.samples}
\alias{get_test_samples.modeler}
\title{Get Test Sample}
\usage{
get_test_samples(obj, ...)

\method{get_test_samples}{samples}(obj)

\method{get_test_samples}{modeler}(obj)
}
\arguments{
\item{obj}{object with valid samples object}

\item{...}{}
}
\value{
test index
}
\description{
Function to extract test sample index from samples object
}
\examples{
# Simple example
library(dplyr)
mtcars \%>\%
add_holdout_samples(., splits = c(.6, .2, .2)) \%>\%
get_test_samples(.)
}
