% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_train_samples}
\alias{get_train_samples}
\alias{get_train_samples.samples}
\alias{get_train_samples.modeler}
\title{Get Training Samples}
\usage{
get_train_samples(obj)

\method{get_train_samples}{samples}(obj)

\method{get_train_samples}{modeler}(obj)
}
\arguments{
\item{obj}{object with valid samples object}

\item{...}{}
}
\value{
list of train indicies
}
\description{
Function to extract training sample indicies from samples object
}
\examples{
# Simple example
library(dplyr)
mtcars \%>\%
add_holdout_samples(., splits = c(.8, .2)) \%>\%
get_train_samples(.)
}
