% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_validation_samples}
\alias{get_validation_samples}
\alias{get_validation_samples.samples}
\alias{get_validation_samples.modeler}
\title{Get Validation Samples}
\usage{
get_validation_samples(obj, ...)

\method{get_validation_samples}{samples}(obj)

\method{get_validation_samples}{modeler}(obj)
}
\arguments{
\item{obj}{object with valid samples object}

\item{...}{}
}
\value{
list of validation indicies
}
\description{
Function to extract validation sample indicies from samples object
}
\examples{
# Simple example
library(dplyr)
mtcars \%>\%
add_holdout_samples(., splits = c(.8, .2)) \%>\%
get_validation_samples(.)
}
