% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_barchart.R
\name{gg_barchart}
\alias{gg_barchart}
\title{GGplot2 Bar Chart}
\usage{
gg_barchart(df, x_variable, y_variable = NULL, weight = NULL,
  fill = sncr_pal()(1), color = "black", position = "dodge",
  proportion = FALSE, sort = FALSE, desc = TRUE, label = FALSE,
  label_args = list(vjust = "bottom"), facet_formula = NULL,
  facet_labeller = "both", facet_args = list(), coord = NULL,
  coord_args = list(), title = NULL, subtitle = NULL,
  x_axis_title = NULL, y_axis_title = NULL, caption = NULL,
  legend_args = list(), theme = "sncr", palette = "a2", ...)
}
\arguments{
\item{x_variable}{x-variable to chart. expects a string. required input}

\item{y_variable}{y-variable to chart. default is NULL. If NULL then the
x-variable is summarized and either a count or proportion is rendered.
expects a string}

\item{weight}{variable to sum the count of the x_variable by. Default is NULL.
expects a string with column name}

\item{fill}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to fill the histogram
with color. expects a string. default is 'grey25'}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{proportion}{TRUE/FALSE. If TRUE, computes percentages, otherwise counts
if y-variable not provided. default is FALSE}

\item{sort}{TRUE/FALSE. If TRUE, orders the bars by either the count in the
case of x-variable only or by the y-variable provided. Default is FALSE}

\item{desc}{logical option to sort boxplots in descending order. Default is
TRUE. Note sort has to be TRUE for desc option to apply}

\item{label}{logical. If TRUE creates a geom_label layer and sets the
label variable to the y-variable}

\item{label_args}{list of arguments passed to the geom_label layer. default
sets the vjust to 'bottom'}

\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{facet_labeller}{takes a string. expecting name of one of label_*
functions. ex - "both" creates a facet label with both the column name and
the value. The two most common examples are 'both' and
'value'\url{http://ggplot2.tidyverse.org/reference/#section-facetting-labels}}

\item{facet_args}{a list of arguments to pass to either the facet_grid or
facet_wrap functions. default is a list with only he labeller parameter set
to both colname and field value.
\url{http://ggplot2.tidyverse.org/reference/facet_grid.html} or
\url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}}

\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{coord_args}{list of arguments to pass to the coord_* function created.
default is empty list.}

\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{legend_args}{list of arguments to pass to guide_legend function to
control the legend formatting
\url{http://ggplot2.tidyverse.org/reference/guide_legend.html}. default is
empty list}

\item{theme}{argument to change the chart theme. expects a ggplot_theme
name. default is "sncr" which calls theme_sncr(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a palette name. default is "a2" for a2 palette}

\item{...}{}

\item{...}{Arguments passed on to \code{ggplot2::geom_bar}
\describe{
  \item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
  \item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}
  \item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
  \item{width}{Bar width. By default, set to 90\% of the resolution of the data.}
  \item{binwidth}{\code{geom_bar} no longer has a binwidth argument - if
you use it you'll get an warning telling to you use
\code{\link{geom_histogram}} instead.}
  \item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
  \item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}
  \item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
  \item{stat}{Override the default connection between \code{geom_bar} and
\code{stat_count}.}
}}
}
\value{
returns a ggplot2 object of a formatted bar chart
}
\description{
This is a wrapper function to create a ggplot2 bar chart. This function
unifies much of the ggplot2 functionality into a single function. Default
parameter values are set throughout to allow for quick interactive charting,
but is also highly customizable to allow for more sophisicated charts.
}
\details{
This function takes any of the geom_bar arguments
\url{http://ggplot2.tidyverse.org/reference/geom_bar.html}
}
\examples{
# Create a data set
d <- mtcars \%>\% mutate(am = as.factor(am), cyl = as.factor(cyl))

# Create series of basic charts
gg_barchart(d, "am")
gg_barchart(d, "am", fill="steelblue", color="grey25")
gg_barchart(d, "am", fill="steelblue", color="grey25", title="Stranger", subtitle="Things")
gg_barchart(d, "am", fill="steelblue", color="grey25", title="Stranger", subtitle="Things", theme="minimal")

A few charts with more features
gg_barchart(d, "am", proportion = T)
gg_barchart(d, "am", fill="cyl")
gg_barchart(d, "am", fill="cyl", position="fill")
gg_barchart(d, "am", fill="cyl", label=T)

# Example of x-variable and y-variable
d1 <- mtcars \%>\% group_by(am = factor(am)) \%>\% summarise(count = n())
gg_barchart(d1, x_variable="am", y_variable="count")
}
\seealso{
geom_label arguments
  \url{http://ggplot2.tidyverse.org/reference/geom_text.html}
}
