% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxplot.R
\name{gg_boxplot}
\alias{gg_boxplot}
\title{GGplot2 Boxplot Chart}
\usage{
gg_boxplot(df, x_variable = NULL, y_variable = NULL, ymin = NULL,
  lower = NULL, middle = NULL, upper = NULL, ymax = NULL,
  fill = sncr_pal()(1), color = "black", alpha = 0.6,
  position = "dodge", sort = FALSE, desc = TRUE, facet_formula = NULL,
  facet_labeller = "both", facet_args = list(), coord = NULL,
  coord_args = list(), title = NULL, subtitle = NULL,
  x_axis_title = NULL, y_axis_title = NULL, caption = NULL,
  legend_args = list(), theme = "sncr", palette = "a2", ...)
}
\arguments{
\item{x_variable}{x-variable to chart. expects a string. default is NULL. not
a required input}

\item{y_variable}{continuous y-variable to chart. expects a string. default
is NULL. either a y-variable is required or boxplot values (ymin, lower,
middle, upper, ymax)}

\item{ymin}{computed min value. default is NULL}

\item{lower}{computed lower value of boxplot. default is NULL}

\item{middle}{computed middle value of boxplot. default is NULL}

\item{upper}{computed upper value of boxplot. default is NULL}

\item{ymax}{computed max value of boxplot. default is NULL}

\item{fill}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to fill the histogram
with color. expects a string. default is 'grey25'}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{sort}{logical option to sort boxplots by y_variable. Default is FALSE}

\item{desc}{logical option to sort boxplots in descending order. Default is
TRUE. Note sort has to be TRUE for desc option to apply}

\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{facet_labeller}{takes a string. expecting name of one of label_*
functions. ex - "both" creates a facet label with both the column name and
the value. The two most common examples are 'both' and
'value'\url{http://ggplot2.tidyverse.org/reference/#section-facetting-labels}}

\item{facet_args}{a list of arguments to pass to either the facet_grid or
facet_wrap functions. default is a list with only he labeller parameter set
to both colname and field value.
\url{http://ggplot2.tidyverse.org/reference/facet_grid.html} or
\url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}}

\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{coord_args}{list of arguments to pass to the coord_* function created.
default is empty list.}

\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{legend_args}{list of arguments to pass to guide_legend function to
control the legend formatting
\url{http://ggplot2.tidyverse.org/reference/guide_legend.html}. default is
empty list}

\item{theme}{argument to change the chart theme. expects a ggplot_theme
function. default is theme_grey(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a vector of either color names or hex
values. default is the Set1 Brewer Palette}

\item{...}{}

\item{...}{Arguments passed on to \code{ggplot2::geom_boxplot}
\describe{
  \item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
  \item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}
  \item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
  \item{outlier.colour}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{outlier.color}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{outlier.fill}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{outlier.shape}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{outlier.size}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{outlier.stroke}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{outlier.alpha}{Default aesthetics for outliers. Set to \code{NULL} to inherit from the
  aesthetics used for the box.

  In the unlikely event you specify both US and UK spellings of colour, the
  US spelling will take precedence.}
  \item{notch}{if \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}
  \item{notchwidth}{for a notched box plot, width of the notch relative to
the body (default 0.5)}
  \item{varwidth}{if \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the \code{weight} aesthetic).}
  \item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
  \item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}
  \item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
  \item{stat}{Use to override the default connection between
\code{geom_boxplot} and \code{stat_boxplot}.}
}}
}
\value{
returns a ggplot2 boxplot object
}
\description{
GGplot2 Boxplot Chart
}
\examples{
#' # Create a data set
d <- mtcars \%>\% mutate(am = as.factor(am), cyl = as.factor(cyl))

# Create series of basic charts
gg_boxplot(d, y_variable = "mpg")
gg_boxplot(d, y_variable = "mpg", x_variable = "am")
gg_boxplot(d, y_variable = "mpg", x_variable = "am", fill="blue")
gg_boxplot(d, y_variable ="mpg", notch=T)

# Create boxplot from values
y <- rnorm(100)
df <- data.frame(
 x = 1,
 y0 = min(y),
 y25 = quantile(y, 0.25),
 y50 = median(y),
 y75 = quantile(y, 0.75),
 y100 = max(y))

gg_boxplot(df, ymin = "y0", lower = "y25", middle = "y50", upper = "y75", ymax = "y100", fill="darkorange")
}
