% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{gg_coord}
\alias{gg_coord}
\title{GGPlot2 Coordinate Component}
\usage{
gg_coord(coord = NULL, ...)
}
\arguments{
\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{...}{takes additional arguments to coord function specified}
}
\description{
Wrapper function for ggplot2 coord_* family of functions. Used internally by
sncr_* functions to handle coordinate arguments
}
\examples{
ggplot(mtcars, aes(x=mpg)) +
geom_histogram() +
gg_coord("flip")
}
\seealso{
ggplot2 coordinate fucntions
  \url{http://ggplot2.tidyverse.org/reference/#section-coordinate-systems}
}
