% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_histogram.R
\name{gg_histogram}
\alias{gg_histogram}
\title{GGplot2 Histogram Chart}
\usage{
gg_histogram(df, x_variable, fill = sncr_pal()(1), color = "black",
  alpha = 0.5, facet_formula = NULL, facet_labeller = "both",
  facet_args = list(), coord = NULL, coord_args = list(), title = NULL,
  subtitle = NULL, x_axis_title = NULL, y_axis_title = NULL,
  caption = NULL, legend_args = list(), theme = "sncr", palette = "a2",
  ...)
}
\arguments{
\item{df}{data.frame object with variable to chart}

\item{x_variable}{variable name to chart. expects a string}

\item{fill}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to fill the histogram
with color. expects a string. default is 'grey25'}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{alpha}{argument to change the colour transparency; It is a numeric in
[0,1]. The default is 0.0.}

\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{facet_labeller}{takes a string. expecting name of one of label_*
functions. ex - "both" creates a facet label with both the column name and
the value. The two most common examples are 'both' and
'value'\url{http://ggplot2.tidyverse.org/reference/#section-facetting-labels}}

\item{facet_args}{a list of arguments to pass to either the facet_grid or
facet_wrap functions. default is a list with only he labeller parameter set
to both colname and field value.
\url{http://ggplot2.tidyverse.org/reference/facet_grid.html} or
\url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}}

\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{coord_args}{list of arguments to pass to the coord_* function created.
default is empty list.}

\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{legend_args}{list of arguments to pass to guide_legend function to
control the legend formatting
\url{http://ggplot2.tidyverse.org/reference/guide_legend.html}. default is
empty list}

\item{theme}{argument to change the chart theme. expects a ggplot_theme name.
default is "sncr" which calls theme_sncr(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a palette name. default is "a2" for a2
palette}

\item{...}{}

\item{...}{Arguments passed on to \code{ggplot2::geom_histogram}
\describe{
  \item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
  \item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}
  \item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
  \item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
  \item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}
  \item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
  \item{binwidth}{The width of the bins. The default is to use \code{bins}
  bins that cover the range of the data. You should always override
  this value, exploring multiple widths to find the best to illustrate the
  stories in your data.

  The bin width of a date variable is the number of days in each time; the
  bin width of a time variable is the number of seconds.}
  \item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30}
  \item{stat}{Use to override the default connection between
\code{geom_histogram}/\code{geom_freqpoly} and \code{stat_bin}.}
}}
}
\value{
returns a ggplot2 object
}
\description{
This is a wrapper function to create a ggplot2 histogram chart. This function
unifies much of the ggplot2 layers into a single function. Default parameter
values are set throughout to allow for quick interactive charting, but is also
highly customizable to allow for more sophisicated charts.
}
\details{
This function takes any of the geom_histogram arguments
\url{http://ggplot2.tidyverse.org/reference/geom_histogram.html}
}
\examples{
# Create a data set
library(dplyr)
d <- mtcars \%>\% mutate(am = as.factor(am))

# Create series of basic charts
gg_histogram(d, "mpg")
gg_histogram(d, "mpg", fill="steelblue", color="grey25")
gg_histogram(d, "mpg", fill="steelblue", color="grey25", title="Crouching Plot", caption = "hidden message")
gg_histogram(d, "mpg", fill="steelblue", color="grey25", title="Crouching Plot", caption = "hidden message", theme="minimal")

# A few charts with additional features
gg_histogram(d, "mpg", fill="am", color="am")
gg_histogram(d, "mpg", fill="am", color="am", bins=7)
gg_histogram(d, "mpg", fill="am", color="am", bins=7, palette = pals::viridis(2), alpha=.75)
gg_histogram(d, "mpg", fill="am", facet_formula = "~am", bins=7)
gg_histogram(d, "mpg", fill="am", facet_formula = "~am", facet_args = list(scales="free"), bins=7)
gg_histogram(d, "mpg", fill="am", facet_formula = "cyl~am", facet_args = list(scales="free_y"), bins=7)
gg_histogram(d, "mpg", fill="am", color="am", coord="flip", bins=7)

}
\seealso{
For more details on the geom_histogram check out
 \url{http://ggplot2.tidyverse.org/reference/geom_histogram.html}
}
