% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{gg_line}
\alias{gg_line}
\title{GGPlot Line Component}
\usage{
gg_line(line = TRUE, fun = "mean", geom = "line", ...)
}
\arguments{
\item{line}{TRUE/FALSE. Default is TRUE. Option to add geom_line layer}

\item{...}{Arguments passed on to \code{ggplot2::stat_summary}
\describe{
  \item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}
  \item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}
  \item{geom}{Use to override the default connection between
\code{geom_histogram}/\code{geom_freqpoly} and \code{stat_bin}.}
  \item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
  \item{fun.data}{A function that is given the complete data and should
return a data frame with variables \code{ymin}, \code{y}, and \code{ymax}.}
  \item{fun.ymin}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}
  \item{fun.y}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}
  \item{fun.ymax}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}
  \item{fun.args}{Optional additional arguments passed on to the functions.}
  \item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
  \item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
  \item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}}
}
\value{
either a ggplot geom_line layer object or NULL
}
\description{
Wrapper function for adding geom_line layer by way of stat_summary
}
\seealso{
\url{http://ggplot2.tidyverse.org/reference/geom_point.html}
}
