% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_line_chart.R
\name{gg_line_chart}
\alias{gg_line_chart}
\title{GGplot2 Line Chart}
\usage{
gg_line_chart(df, x_variable, y_variable, position = "identity",
  color = sncr_pal()(1), size = 1, alpha = 1, linetype = 1,
  group = NULL, points = F, point_args = list(shape = 16), label = F,
  label_args = list(vjust = "bottom"), facet_formula = NULL,
  facet_labeller = "both", facet_args = list(), coord = NULL,
  coord_args = list(), title = NULL, subtitle = NULL,
  x_axis_title = NULL, y_axis_title = NULL, caption = NULL,
  legend_args = list(), theme = "sncr", palette = "a2", ...)
}
\arguments{
\item{df}{data.frame object with variable to chart}

\item{x_variable}{x-variable to chart. expects a string. required input}

\item{y_variable}{y-variable to chart. expects a string. required input}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{group}{a variable name to apply a group by operation and create
subgroups of line charts}

\item{points}{TRUE/FALSE. If TRUE scatter plot points are added to the chart}

\item{point_args}{list of arguments to pass to geom_point function. default
is an empty list}

\item{label}{logical. If TRUE creates a geom_label layer and sets the
label variable to the y-variable}

\item{label_args}{list of arguments passed to the geom_label layer. default
sets the vjust to 'bottom'}

\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{facet_labeller}{takes a string. expecting name of one of label_*
functions. ex - "both" creates a facet label with both the column name and
the value. The two most common examples are 'both' and
'value'\url{http://ggplot2.tidyverse.org/reference/#section-facetting-labels}}

\item{facet_args}{a list of arguments to pass to either the facet_grid or
facet_wrap functions. default is a list with only he labeller parameter set
to both colname and field value.
\url{http://ggplot2.tidyverse.org/reference/facet_grid.html} or
\url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}}

\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{coord_args}{list of arguments to pass to the coord_* function created.
default is empty list.}

\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{legend_args}{list of arguments to pass to guide_legend function to
control the legend formatting
\url{http://ggplot2.tidyverse.org/reference/guide_legend.html}. default is
empty list}

\item{theme}{argument to change the chart theme. expects a ggplot_theme
name. default is "sncr" which calls theme_sncr(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a palette name. default is "a2" for a2 palette}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
returns a ggplot2 object of a formatted bar chart
}
\description{
This is a wrapper function to create a ggplot2 line chart. This function
unifies much of the ggplot2 functionality into a single function. Default
parameter values are set throughout to allow for quick interactive charting,
but is also highly customizable to allow for more sophisicated charts. There
is an option to add points to the line as well.
}
\details{
This function takes any of the geom_line arguments
\url{http://ggplot2.tidyverse.org/reference/geom_line.html}
}
\examples{
library(ggplot2)
library(dplyr)
d <- mtcars \%>\% mutate(am = as.factor(am))
gg_line_chart(d, "wt", "mpg", facet_formula = ".~cyl", facet_args = list(scales = "free"), color = "blue")
gg_line_chart(d, "wt", "mpg")
gg_line_chart(d, "wt", "mpg", point = T)
gg_line_chart(d, "wt", "mpg", point = T, point_args = list(size = 3))
gg_line_chart(d, "wt", "mpg", color = "red", linetype = "dotdash")
gg_line_chart(d, "wt", "mpg", color = "red", point = T, point_args = list(size = 3))
gg_line_chart(d, "wt", "mpg", color = 'am', point = T, point_args = list(size = 3, shape = 15))
gg_line_chart(d, "wt", "mpg", point = T, point_args = list(size = 3, shape = 15, color = 'red'))
}
\seealso{
geom_label arguments
  \url{http://ggplot2.tidyverse.org/reference/geom_text.html}
}
