% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatter_chart.R
\name{gg_scatter_chart}
\alias{gg_scatter_chart}
\title{GGplot Smoother Chart}
\usage{
gg_scatter_chart(df, x_variable, y_variable, color = "grey25", size = 2,
  alpha = 1, shape = 16, points = TRUE, position = "identity",
  smoother = FALSE, smooth_method = "auto", smooth_args = list(),
  smooth_formula = as.formula("y~x"), smooth_ci = TRUE,
  smooth_color = "red", smooth_fill = "grey75", facet_formula = NULL,
  facet_labeller = "both", facet_args = list(), coord = NULL,
  coord_args = list(), title = NULL, subtitle = NULL,
  x_axis_title = NULL, y_axis_title = NULL, caption = NULL,
  legend_args = list(), theme = "sncr", palette = "a2", ...)
}
\arguments{
\item{df}{data.frame object with variable to chart}

\item{x_variable}{x-variable to chart. expects a string. required input}

\item{y_variable}{y-variable to chart. expects a string. required input}

\item{color}{either a variable name to apply a group by operation and create
subgroups of histograms or color name or hex value to color the outlines of
the histogram. expects a string. default is 'black'}

\item{points}{TRUE/FALSE. If TRUE scatter plot points are added to the chart}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{smooth_method}{string name for smoother method to apply. accepts "lm",
"glm", "rlm", "loess", "gam", "rlm"}

\item{smooth_args}{list of additional arguments to pass to method}

\item{smooth_formula}{regression formula provided to smoother method. ex -
y~x}

\item{smooth_ci}{logical option to add smoother confidence interval to chart}

\item{smooth_color}{color option for smoother line}

\item{smooth_fill}{fill option for smoother confidence interval}

\item{facet_formula}{a valid r formula comprised of column names to create
facets. expects a string. formula can either be a '~rhs' formula for a
facet_wrap or 'lhs~rhs' to create a facet_grid. Default is NULL which
creates no facets}

\item{facet_labeller}{takes a string. expecting name of one of label_*
functions. ex - "both" creates a facet label with both the column name and
the value. The two most common examples are 'both' and
'value'\url{http://ggplot2.tidyverse.org/reference/#section-facetting-labels}}

\item{facet_args}{a list of arguments to pass to either the facet_grid or
facet_wrap functions. default is a list with only he labeller parameter set
to both colname and field value.
\url{http://ggplot2.tidyverse.org/reference/facet_grid.html} or
\url{http://ggplot2.tidyverse.org/reference/facet_wrap.html}}

\item{coord}{name of coordinate transform. ex - 'flip' creates a coord_flip
and a 'trans' creates a coord_trans. expects a string. default is NULL with
creates no coordinate changes}

\item{coord_args}{list of arguments to pass to the coord_* function created.
default is empty list.}

\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}

\item{legend_args}{list of arguments to pass to guide_legend function to
control the legend formatting
\url{http://ggplot2.tidyverse.org/reference/guide_legend.html}. default is
empty list}

\item{theme}{argument to change the chart theme. expects a ggplot_theme
name. default is "sncr" which calls theme_sncr(). check out
\url{http://ggplot2.tidyverse.org/reference/#section-themes} for more
options}

\item{palette}{argument to change the color palette used is either the fill or
color is set to a variable. expects a palette name. default is "a2" for a2 palette}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{smooth}{logical option to add smoother to chart}
}
\value{
returns a ggplot2 geom_smooth object
}
\description{
This is a wrapper function to create a ggplot2 xy scatter and smoother chart.
This function unifies much of the ggplot2 functionality into a single
function. Default parameter values are set throughout to allow for quick
interactive charting, but is also highly customizable to allow for more
sophisicated charts.
}
\details{
This function combines many of the features from the ggplot2 function's:
geom_point \url{http://ggplot2.tidyverse.org/reference/geom_point.html} and
geom_smooth  \url{http://ggplot2.tidyverse.org/reference/geom_smooth.html}
}
\examples{
# Create a data set
d <- mtcars \%>\% mutate(am = as.factor(am), cyl = as.factor(cyl))

# Create series of basic charts
gg_scatter_chart(d, y_variable = "mpg", x_variable = "wt")
gg_scatter_chart(d, y_variable = "mpg", x_variable = "wt")
gg_scatter_chart(d, y_variable = "mpg", x_variable = "wt", color="wt")
gg_scatter_chart(d, y_variable ="mpg", x_variable = "wt", points=F)

# Smoother Options
gg_scatter_chart(d, y_variable = "mpg", x_variable = "wt", smoother = TRUE, smooth_method="lm") # linear model
gg_scatter_chart(d, y_variable = "mpg", x_variable = "wt", smoother = TRUE, smooth_method="lm", smooth_ci=F, smooth_color="red")
gg_scatter_chart(d, y_variable = "mpg", x_variable = "wt", smoother = TRUE, smooth_method="lm", smooth_ci=F, smooth_color="red", smooth_formula=y~splines::bs(x, degree=1, knots=2)) # linear spline model
gg_scatter_chart(d, y_variable = "vs", x_variable = "wt", smoother = TRUE, smooth_method="glm", smooth_color="red", smooth_args=list(family = "binomial")) # binomial logistic model

}
\seealso{
geom_point
  \url{http://ggplot2.tidyverse.org/reference/geom_point.html}

geom_smooth
  \url{http://ggplot2.tidyverse.org/reference/geom_smooth.html}
}
