% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{gg_sort}
\alias{gg_sort}
\title{GGPlot Sort Component}
\usage{
gg_sort(sort, x_variable, y_variable, x_axis_title)
}
\value{
list with sort params arguments and updated x axis title
}
\description{
Wrapper function to sort x_variable by either count or median of y-variable
if provided
}
\seealso{
\code{\link[forcats]{fct_infreq}}

\code{\link[forcats]{fct_reorder}}
}
