% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_components.R
\name{gg_titles}
\alias{gg_titles}
\title{GG Titles Component}
\usage{
gg_titles(title, subtitle, x_axis_title, y_axis_title, caption)
}
\arguments{
\item{title}{chart title. expects a string}

\item{subtitle}{chart subtitle. expects a string}

\item{x_axis_title}{x-axis title. expects a string}

\item{y_axis_title}{y-axis title. expects a string}

\item{caption}{chart caption. expects a string}
}
\value{
ggplot2 label object
}
\description{
Wrapper function to labs function in ggplot2. Needed to handle parameter
inputs to the title values with default \code{link{waiver}}
}
