% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter_visuals.R
\name{hc_detect_anomalies_chart}
\alias{hc_detect_anomalies_chart}
\title{Highcharts Anomaly Chart}
\usage{
hc_detect_anomalies_chart(df, title = "Anomaly Detecter",
  subtitle = "Anomaly Chart", credits = "Synchronoss")
}
\description{
Highcharts Anomaly Chart
}
