% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecter_visuals.R
\name{hc_detect_decomp_chart}
\alias{hc_detect_decomp_chart}
\title{Highcharts Detecter Decomposition Chart}
\usage{
hc_detect_decomp_chart(df, title = "Anomaly Detecter",
  subtitle = "Decomposition Chart", credits = "Synchronoss")
}
\description{
Highcharts Detecter Decomposition Chart
}
