% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster_visuals.R
\name{hc_forecast_chart}
\alias{hc_forecast_chart}
\title{Highcharts Forecast Chart}
\usage{
hc_forecast_chart(fobj, title = "Forecaster", subtitle = "Forecast Chart",
  credits = "Synchronoss")
}
\description{
Highcharts Forecast Chart
}
