% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute_constant_at_chr}
\alias{impute_constant_at_chr}
\title{Impute Character Constant At}
\usage{
impute_constant_at_chr(df, measure_vars, fill)
}
\description{
Function imputes character constants for numeric columns
}
