% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute_constant_at_num}
\alias{impute_constant_at_num}
\title{Impute Numeric Constant At}
\usage{
impute_constant_at_num(df, measure_vars, fill)
}
\description{
Function imputes numeric constants for numeric columns
}
