% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute_mode}
\alias{impute_mode}
\title{Impute Mode}
\usage{
impute_mode(df, measure_vars = NULL)
}
\arguments{
\item{df}{Spark or R data.frame}

\item{measure_vars}{optional character vector of column names to apply impute
function on. default is NULL which applies impute on all columns with
missing values}
}
\value{
imputed Spark or R data.frame
}
\description{
Function to impute missing values with the most frequent values (mode). Applied columwise -
the most frequent value imputed is the by column
}
\details{
impute mode does not support grouping currently
}
