% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mape}
\alias{mape}
\alias{mape.data.frame}
\alias{mape.tbl_spark}
\title{Generic mape function}
\usage{
mape(...)

\method{mape}{data.frame}(x, predicted, actual)

\method{mape}{tbl_spark}(x, predicted, actual)
}
\description{
Generic mape function
}
