% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Model Class Helper}
\usage{
model(pipeline = pipeline(), method, ..., desc = NULL, path = NULL)
}
\arguments{
\item{pipeline}{pipeline object. default is empty pipeline which applies no
data transformations}

\item{method}{string input of model method}

\item{...}{additional arguments to pass to model method}

\item{desc}{optional model description}

\item{path}{optional file path to save model}
}
\description{
Function to create a model object.
}
\details{
Requires a pipeline object input, and a valid model method. Any method
package dependencies need to be loaded prior to model call
}
