% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{model_methods}
\alias{model_methods}
\title{Available Model Methods for Modeler Objects}
\format{A data frame with 24 rows and 5 variables:
 \describe{
  \item{type}{modeler type}
  \item{method}{model method function name}
  \item{name}{name of method}
  \item{package}{R package backend}
  \item{class}{method class membership. can be more than 1} ...
  }}
\usage{
model_methods
}
\description{
A dataset containing the avaialble model methods for each modeler type
}
\keyword{datasets}
