% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{new_forecast_fit_args}
\alias{new_forecast_fit_args}
\title{Forecast Fit Arguments Constructer function}
\usage{
new_forecast_fit_args(model, df, index_var, y_var, x_vars, boxcox, ...)
}
\description{
Forecast Fit Arguments Constructer function
}
