% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{new_forecaster}
\alias{new_forecaster}
\alias{forecaster}
\title{Forecaster Constructer Function}
\usage{
new_forecaster(df, target, index_var, unit = NULL, frequency = NULL,
  prediction_conf_levels = c(80, 95), name = NULL, id = NULL,
  version = NULL, desc = NULL, scientist = NULL, dir = NULL, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{target}{target variable for supervised use cases such as forecaster,
regressor, and classifier}

\item{index_var}{column name of index variable. index variable should be
either sequential numeric, date or datetime.}

\item{frequency}{seasonaly frequency of target. Default is NULL - no
seasonality}

\item{prediction_conf_levels}{prediction confidence levels. Default is 80 and
95 percent CIs}

\item{name}{optional name for use case. Default is to append target with type}

\item{id}{optional id string. A random id will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{dir}{optional directory path for saving reports, data and models}

\item{...}{any addtional parameters}

\item{units}{index variable unit for date or datetime variables. ex - days
for date index variable.}
}
\description{
Creates a forecaster object which inherits from modeler class
}
\seealso{
Other use cases: \code{\link{modeler}},
  \code{\link{new_classifier}},
  \code{\link{new_regressor}}, \code{\link{new_segmenter}}
}
