% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{new_forecaster_fit_args}
\alias{new_forecaster_fit_args}
\title{Forecast Fit Arguments Constructer function}
\usage{
new_forecaster_fit_args(algo, df, index_var, y_var, x_vars, boxcox, name, ...)
}
\description{
Forecast Fit Arguments Constructer function
}
