% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{new_forecaster_method}
\alias{new_forecaster_method}
\title{Forecast Method Constructer function}
\usage{
new_forecaster_method(algo, df, index_var, y_var, x_vars, boxcox, name, ...)
}
\description{
Forecast Method Constructer function
}
