% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{new_model}
\alias{new_model}
\title{Model Class Constructer}
\usage{
new_model(pipe, target, method, method_args, class, desc, path, id, status,
  created_on, last_updated, fit, performance)
}
\description{
Model Class Constructer
}
