% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{new_modeler}
\alias{new_modeler}
\title{New Modeler Object Constructer function}
\usage{
new_modeler(df, target, type, name, id, version, desc, scientist, ...)
}
\description{
New Modeler Object Constructer function
}
