% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{new_roller_args}
\alias{new_roller_args}
\title{Roller Arguments Constructor function}
\usage{
new_roller_args(order_vars, group_vars, measure_vars, fun, width, by, partial,
  mode, ...)
}
\value{
roller_args object
}
\description{
Creates new object of class roller_args
}
