% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenter.R
\name{new_segmenter}
\alias{new_segmenter}
\alias{segmenter}
\title{Segmenter Constructer Function}
\usage{
new_segmenter(df, name = NULL, id = NULL, version = NULL, desc = NULL,
  scientist = NULL, dir = NULL, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{name}{optional name for use case. Default is to append target with type}

\item{id}{optional id string. A random id will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{dir}{optional directory path for saving reports, data and models}

\item{...}{any addtional parameters}
}
\description{
Creates a segmenter object which inherits from modeler class
}
\seealso{
Other use cases: \code{\link{modeler}},
  \code{\link{new_forecaster}}
}
