% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{new_summariser_args}
\alias{new_summariser_args}
\title{Summariser Constructor Function}
\usage{
new_summariser_args(group_vars, measure_vars, fun, ...)
}
\arguments{
\item{group_vars}{vector of column names. used in the group by aggregration.
can be zero or more columns. set to NULL for no grouping}

\item{measure_vars}{vector of column names. colnames to aggregrate. can be one
or more columns. Default is NULL which will calculate counts only}

\item{fun}{vector of function names. can be one or more columns. default is
sum. Current functions supported: n_distinct, min, max, sum, mean, variance,
sd, kurtosis, skewness, and percentile}

\item{...}{additional parameters passed to aggregate function}
}
\description{
Function to create object of summarise_args S3 class
}
