% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{normalizer}
\alias{normalizer}
\alias{normalizer.data.frame}
\alias{normalizer.tbl_spark}
\title{DataFrame Normalizer Data Transform Function}
\usage{
normalizer(df, ...)

\method{normalizer}{data.frame}(df, group_vars = NULL, measure_vars)

\method{normalizer}{tbl_spark}(df, group_vars = NULL, measure_vars)
}
\arguments{
\item{df}{DataFrame}

\item{group_vars}{optional vector of column names. Applies transformation on
grouped data. Can be zero or more columns. Default set to NULL (no grouping)}

\item{measure_vars}{one or more variable names. vector of strings. columns to
apply transformation on}
}
\value{
Transformed Dataframe with the same dimensions of input dataframe
}
\description{
Data transformation function to normalize one or more variables. The normalize
function transforms values by subtracting the min and dividing by the max of
all values. The function is applied columnwise only. The function only applies
to numerical data types.
}
\details{
The normalize function transforms values to a [0,1] range. The underlying
distribution is perserved.
}
\examples{
library(dplyr)
normalizer(mtcars, measure_vars = c("mpg", "hp"))
normalizer(mtcars, group_vars = "am", measure_vars = c("mpg", "hp"))
}
