% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline Class Helper}
\usage{
pipeline(expr = identity, desc = NULL)
}
\description{
Creates a pipeline object. The expression is set by default as the identity function. This returns the
}
