% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/forecaster.R, R/segmenter.R,
%   R/spark-ml.R
\name{predict.forecast_model}
\alias{predict.forecast_model}
\alias{predict.forecaster}
\alias{predict.segmenter}
\alias{predict.spark_ml}
\title{Forecast Prediction Method}
\usage{
\method{predict}{forecast_model}(obj, periods, data = NULL, level = c(80,
  95))

\method{predict}{forecaster}(obj, periods, data = NULL, level = c(80, 95),
  desc = "")

\method{predict}{segmenter}(obj, data = NULL, desc = "", ...)

\method{predict}{spark_ml}(obj, data, ...)
}
\description{
Forecast Prediction Method

Method makes predictions for Forecaster's final model

Method makes predictions using segmenter object's final model

Predict Method for Spark-ML Object
}
