% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{rmse}
\alias{rmse}
\alias{rmse.data.frame}
\alias{rmse.tbl_spark}
\title{Generic rmse function}
\usage{
rmse(...)

\method{rmse}{data.frame}(x, predicted, actual)

\method{rmse}{tbl_spark}(x, predicted, actual)
}
\description{
Generic rmse function

rmse.tbl_spark <- function(x, predicted, actual){
  checkmate::assert_choice(predicted, colnames(x))
  checkmate::assert_choice(actual, colnames(x))
  x %>%
    dplyr::summarise_at(predicted, funs(rmse = sqrt(mean((. - !!rlang::sym(actual))^2)))) %>%
    dplyr::collect()
}
}
