% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rsample}
\alias{rsample}
\alias{rsample.data.frame}
\alias{rsample.numeric}
\alias{rsample.tbl_spark}
\title{Random Sample function}
\usage{
rsample(x, ...)

\method{rsample}{data.frame}(df, amount, seed = NULL)

\method{rsample}{numeric}(x, amount, seed = NULL)

\method{rsample}{tbl_spark}(df, amount, seed = NULL)
}
\arguments{
\item{amount}{percent of data to randomly sample}

\item{seed}{optional input for setting random seed}
}
\description{
Function creates a random sample index from a dataset. Index can be used to
create data samples
}
