% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{samples}
\alias{samples}
\title{Sample Class Helper function}
\usage{
samples(validation_method, validation_args, test_holdout_prct,
  test_holdout_method, downsample_prct, train_indicies, validation_indicies,
  indicies_names, test_index)
}
\description{
Sample Class Helper function
}
