% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{set_final_model}
\alias{set_final_model}
\alias{set_final_model.modeler}
\title{Set Final Model Generic}
\usage{
set_final_model(obj, method, id, reevaluate, refit)

\method{set_final_model}{modeler}(obj, method, id = NULL, reevaluate = TRUE,
  refit = TRUE)
}
\arguments{
\item{obj}{forecaster object}

\item{method}{selection method. Choices are 'manual' which requires a valid
id or 'best' method which selects the best model based on the measure}

\item{id}{optional model id input required for manual method}

\item{reevaluate}{logical option to re-evaluate the final model. Requires a
test holdout sample}

\item{refit}{logical option to re-fit final model on entire training dataset}
}
\description{
Function to select the final model. See method argument for options
}
