% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{set_sampling.modeler}
\alias{set_sampling.modeler}
\title{Set Modeler Sampling Methods}
\usage{
\method{set_sampling}{modeler}(mobj, validation_method, validation_args,
  test_method, test_args)
}
\arguments{
\item{mobj}{modeler object}

\item{validation_method}{sampling method to apply to create training and validation datasets}

\item{validation_args}{list of additional arguments to pass to validation sampling method}

\item{test_method}{sampling method to apply to create test dataset only. default is 'none'}

\item{test_args}{list of additional arguments to pass to test sampling method}
}
\description{
Function sets internal Modeler object sampling methods. Function creates sample indicies but not
full samples. Modeler object sampling slots are updated
}
