% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{silhouette}
\alias{silhouette}
\alias{silhouette.tbl_spark}
\title{Generic silhouette function}
\usage{
silhouette(...)

\method{silhouette}{tbl_spark}(x, predicted = "predicted")
}
\description{
Generic silhouette function
}
