% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{sip_add_dataset}
\alias{sip_add_dataset}
\title{SIP Metastore Add Dataset}
\usage{
sip_add_dataset(dataset_name, dataset_path, dataset_format, row_count = NULL,
  component, desc, project_id, batch_id, input_datasets_id, started, finished,
  status = "SUCCESS", hostname)
}
\arguments{
\item{dataset_name}{name of dataset to add}

\item{dataset_path}{file path to saved dataset}

\item{dataset_format}{format of dataset. accepts parquet, JSON, csv and rds}

\item{row_count}{dataset record count. Default is NULL}

\item{component}{name of component used to create dataset}

\item{desc}{description of dataset}

\item{project_id}{project id string}

\item{batch_id}{batch id string}

\item{input_datasets_id}{string vector of input datasets}

\item{started}{string with starting time. Format should be YYYYMMDD-HHMMSS}

\item{finished}{string with starting time. Format should be YYYYMMDD-HHMMSS}

\item{status}{component status. default is 'SUCCESS'}

\item{hostname}{hostname only of saw security location. No pathing required}
}
\value{
POST response content
}
\description{
Adds a dataset record to the SIP Metastore
}
\details{
POST request to /saw/services/internal/workbench/projects/{project_id}/datasets/create
}
