% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{sip_get_datapath}
\alias{sip_get_datapath}
\title{SIP Metastore get Data File Path}
\usage{
sip_get_datapath(name, project_id, hostname, token)
}
\arguments{
\item{name}{dataset name}

\item{project_id}{id of project to retrieve datasets from}

\item{hostname}{hostname only of saw security location. No pathing required
Ex: hostname = "https://saw-bda-cert-vaste.sncrcorp.net"}

\item{token}{valid SIP token. Result of sip_authenticate function}
}
\value{
dataset physical file path location string
}
\description{
Function to call SIP Rest API to generate valid file path for dataset
}
\details{
API handles the project
}
