% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{sip_get_dataset_details}
\alias{sip_get_dataset_details}
\title{SIP Metastore Get Dataset Details}
\usage{
sip_get_dataset_details(dataset_name, project_id, hostname, token)
}
\arguments{
\item{dataset_name}{dataset name}

\item{project_id}{id of project to retrieve datasets from}

\item{hostname}{hostname only of saw security location. No pathing required
Ex: hostname = "https://saw-bda-cert-vaste.sncrcorp.net"}

\item{token}{valid SIP token. Result of sip_authenticate function}
}
\value{
list with dataset details
}
\description{
Returns dataset metadata details
}
\details{
GET request to saw/services/internal/workbench/projects/{project_id}/datasets/{dataset_id}
}
