% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-palette-theme.R
\name{sncr_cols}
\alias{sncr_cols}
\title{Function to extract Synchronoss colors as hex codes}
\usage{
sncr_cols(...)
}
\arguments{
\item{...}{Character names of sncr_colors}
}
\description{
Function to extract Synchronoss colors as hex codes
}
