% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{sql_over_translator}
\alias{sql_over_translator}
\title{OVER SQL Translation funcion}
\usage{
sql_over_translator(var, fun, group_vars, order_vars, width, new_name = NULL)
}
\description{
Helper function to create OVER sql statement
}
