% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{standardizer}
\alias{standardizer}
\alias{standardizer.data.frame}
\alias{standardizer.tbl_spark}
\title{DataFrame Standarizer Data Transform Function}
\usage{
standardizer(df, ...)

\method{standardizer}{data.frame}(df, group_vars = NULL, measure_vars)

\method{standardizer}{tbl_spark}(df, group_vars = NULL, measure_vars)
}
\arguments{
\item{df}{DataFrame}

\item{group_vars}{optional vector of column names. Applies transformation on
grouped data. Can be zero or more columns. Default set to NULL (no grouping)}

\item{measure_vars}{one or more variable names. vector of strings. columns to
apply transformation on}
}
\value{
Transformed Dataframe with the same dimensions of input dataframe
}
\description{
Data transformation function to standarize one or more variables. The
standardize function transforms values by subtracting the mean and dividing by
the standard deviation of all values. The function is applied columnwise only.
The function only applies to numerical data types.
}
\details{
The standardize function transforms values to be on a standard normal
distribution N(0, 1) scale [-Inf, Inf].
}
\examples{
library(dplyr)
standardizer(mtcars, measure_vars = c("mpg", "hp"))
standardizer(mtcars, group_vars = "am", measure_vars = c("mpg", "hp"))
}
