% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{summariser}
\alias{summariser}
\alias{summariser.data.frame}
\alias{summariser.tbl_spark}
\title{DataFrame Summariser function}
\usage{
summariser(df, ...)

\method{summariser}{data.frame}(df, group_vars, measure_vars = NULL,
  fun = c("sum"), ...)

\method{summariser}{tbl_spark}(df, group_vars, measure_vars = NULL,
  fun = c("sum"), ...)
}
\arguments{
\item{df}{DataFrame}

\item{...}{additional parameters passed to aggregate function}

\item{group_vars}{vector of column names. used in the group by aggregration.
can be zero or more columns. set to NULL for no grouping}

\item{measure_vars}{vector of column names. colnames to aggregrate. can be one
or more columns. Default is NULL which will calculate counts only}

\item{fun}{vector of function names. can be one or more columns. default is
sum. Current functions supported: n_distinct, min, max, sum, mean, variance,
sd, kurtosis, skewness, and percentile}
}
\value{
returns DataFrame with one row for each unique combination of grouping
 variables provided
}
\description{
Aggregation function. Configuarable inputs for grouping variables, measure
variables and aggregation functions.
}
\examples{
library(dplyr)


# Create toy dataset
set.seed(319)
id_vars <- seq(101, 200, by=1)
dates <- seq(from=Sys.Date()-365, to=Sys.Date(), by="day")
cat1 <- c("A", "B")
cat2 <- c("X", "Y", "Z")

dat <- data.frame()
for(id in id_vars){
  n <- floor(runif(1)*100)
  d <- data.frame(id = id,
                 date = sample(dates, n, replace = TRUE),
                 cat1 = sample(cat1, n, replace = TRUE),
                 cat2 = sample(cat2, n, replace = TRUE),
                 metric1 = sample(1:5, n, replace = TRUE),
                 metric2 = rnorm(n, mean=50, sd = 5))
 dat <- rbind(dat, d)
}

summariser(dat,
          group_vars = c("id", "date", "cat1", "cat2"),
          measure_vars = c("metric1", "metric2"),
          fun = c("sum"))
}
