% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R, R/summariser.R
\name{summariser_args}
\alias{summariser_args}
\title{Roller Argument Helper Function}
\usage{
summariser_args(group_vars, measure_vars, fun, ...)

summariser_args(group_vars, measure_vars, fun, ...)
}
\arguments{
\item{group_vars}{vector of column names. used in the group by aggregration.
can be zero or more columns. set to NULL for no grouping}

\item{measure_vars}{vector of column names. colnames to aggregrate. can be one
or more columns. Default is NULL which will calculate counts only}

\item{fun}{vector of function names. can be one or more columns. default is
sum. Current functions supported: n_distinct, min, max, sum, mean, variance,
sd, kurtosis, skewness, and percentile}

\item{...}{additional parameters passed to aggregate function}
}
\description{
Creates a valid object of roller_args class

Creates a valid object of summarise_args class
}
\details{
Function should be used in roller internals

Function should be used to pass nested arguments to the summarise_map, map
argument
}
