% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{test}
\alias{test}
\alias{test.pipeline}
\title{Pipeline Test function}
\usage{
test(x, pipe, n)

\method{test}{pipeline}(x, pipe, n = 100)
}
\arguments{
\item{x}{input object to apply pipeline expression on. Can either be
data.frame, spark dataframe or modeler object}

\item{pipe}{pipeline object}

\item{n}{number of records to take from head of input. default is 100}
}
\value{
data output resulting from sampled pipeline execution
}
\description{
Test function applies pipeline expression on sample of data only.
}
\examples{

pipe <- pipeline(expr = function(e) mean(e$mpg))
flow(mtcars, pipe)
}
