% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/modeler.R
\name{tidy_performance}
\alias{tidy_performance}
\alias{tidy_performance}
\title{Return Tidy Dataset of Model Performance}
\usage{
tidy_performance(obj)

tidy_performance(obj)
}
\description{
Return Tidy Dataset of Model Performance

Function to extract model performance on all samples. Converts output to
data.frame
}
