% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R, R/spark-ml.R
\name{train.forecast_model}
\alias{train.forecast_model}
\alias{train}
\alias{train.spark_ml}
\title{Train Model Generic}
\usage{
\method{train}{forecast_model}(mobj, indicies, level)

train(...)

\method{train}{spark_ml}(mobj, indicies, ...)
}
\value{
updated model object
}
\description{
Train single model to indicies provided

Train Method for Spark-ML Object
}
\details{
Fits Model and makes predictions for any validation or test indicies
provided. Adds fitted values and predictions to model's performance values
}
