% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R, R/modeler.R, R/segmenter.R
\name{train_models.forecaster}
\alias{train_models.forecaster}
\alias{train_models}
\alias{train_models.segmenter}
\title{Train Model Generic Function}
\usage{
\method{train_models}{forecaster}(obj, ids = NULL)

train_models(obj, ...)

\method{train_models}{segmenter}(obj, ids = NULL)
}
\value{
updated modeler object
}
\description{
Train all models added to a modeler object
}
\details{
Function looks for all untrained models added to a modeler project. Function
fits model based on modeler samples, model method and model method args.
Function saves model fit, all fitted predictions to the training sample and
any validation predictions
}
