% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{valid_forecast_fit_args}
\alias{valid_forecast_fit_args}
\title{Forecast Fit Arguments Validation function}
\usage{
valid_forecast_fit_args(x)
}
\description{
Forecast Fit Arguments Validation function
}
