% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{validate_roller_args}
\alias{validate_roller_args}
\title{Roller Arguments Validation Function}
\usage{
validate_roller_args(x)
}
\arguments{
\item{x}{obj of class summariser_args}
}
\description{
Checks for valid inputs to roller_args class
}
