% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{validate_summariser_args}
\alias{validate_summariser_args}
\title{Summariser Validation Function}
\usage{
validate_summariser_args(x)
}
\arguments{
\item{x}{obj of class summariser_args}
}
\description{
Checks for valid inputs to summarise_args class
}
