% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_holdout_samples}
\alias{add_holdout_samples}
\alias{add_holdout_samples.integer}
\alias{add_holdout_samples.data.frame}
\alias{add_holdout_samples.tbl_spark}
\alias{add_holdout_samples.modeler}
\title{Add Holdout Samples function}
\usage{
add_holdout_samples(x, ...)

\method{add_holdout_samples}{integer}(x, splits)

\method{add_holdout_samples}{data.frame}(df, splits)

\method{add_holdout_samples}{tbl_spark}(df, splits)

\method{add_holdout_samples}{modeler}(obj, splits)
}
\arguments{
\item{x}{numeric vector}

\item{splits}{numeric vector of holdout splits. Values need to sum to 1. Each
value represents the amount of data partitioned. Order matters - the splits
correspond to train/validation/test*. Test value optional.}

\item{df}{R data.frame}

\item{obj}{modeler object}

\item{df}{Spark Dataframe}
}
\description{
Function creates a new samples object with holdout samples based on
configuration
}
\details{
Holdout samples take sequential row indicies. Holdout samples are used
primarily in forecasting applications but can be used for other modeling
applications

Modeler method returns updated modeler object with new samples object
appendeed. Data.frame method returns samples object
}
\examples{

# Data.frame example
add_holdout_samples(mtcars, splits = c(.8, .2))
}
