% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/spark-ml.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.model}
\alias{evaluate.spark_ml_clustering}
\title{Evaluate Model Generic}
\usage{
evaluate(mobj, measure)

\method{evaluate}{model}(mobj, measure)

\method{evaluate}{spark_ml_clustering}(mobj, measure)
}
\arguments{
\item{mobj}{model object}

\item{measure}{measure object}

\item{target_df}{dataframe with target and index variables}
}
\value{
returns evaluted model object
}
\description{
Function to evaluate the predictive performance of a model

Evaluate Method for Spark-ML Clustering Object
}
