% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R, R/segmenter.R
\name{evaluate_models}
\alias{evaluate_models}
\alias{evaluate_models.modeler}
\alias{evaluate_models.segmenter}
\title{Evalutate Models Generic Function}
\usage{
evaluate_models(obj, ...)

\method{evaluate_models}{modeler}(obj, ids = NULL)

\method{evaluate_models}{segmenter}(obj, ids = NULL)
}
\value{
updated modeler object
}
\description{
Evaluate the accuracy of all trained models.
}
\details{
Function applies the measure function associated with modeler object to all
model sample predictions.
}
