% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R, R/modeler.R
\name{evaluate_models.forecaster}
\alias{evaluate_models.forecaster}
\alias{evaluate_models}
\title{Evalutate Models Generic Function}
\usage{
\method{evaluate_models}{forecaster}(obj, ids = NULL)

evaluate_models(obj, ...)
}
\value{
updated modeler object
}
\description{
Evaluate the accuracy of all trained models.
}
\details{
Function applies the measure function associated with modeler object to all
model sample predictions.
}
