% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R, R/modeler.R
\name{evaluate_models.forecaster}
\alias{evaluate_models.forecaster}
\alias{evaluate_models}
\title{Evalutate Models Generic}
\usage{
\method{evaluate_models}{forecaster}(obj)

evaluate_models(obj, ...)
}
\description{
Evalutate Models Generic
}
