% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R, R/spark-ml.R
\name{fit.forecast_model}
\alias{fit.forecast_model}
\alias{fit}
\alias{fit.spark_ml}
\title{Fit Model Generic}
\usage{
\method{fit}{forecast_model}(obj, data, ...)

fit(...)

\method{fit}{spark_ml}(obj, data)
}
\description{
Fit model to single data sample

Spark ML fit method
}
