% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{fit.forecast_model}
\alias{fit.forecast_model}
\alias{fit}
\title{Fit Model Generic}
\usage{
\method{fit}{forecast_model}(obj, data, ...)

fit(...)
}
\description{
Fit model to single data sample
}
