% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{fit_args.forecaster}
\alias{fit_args.forecaster}
\title{Forecaster Fit Arguments Helper function}
\usage{
\method{fit_args}{forecaster}(algo = "auto.arima", df, index_var, y_var,
  x_vars = NULL, boxcox = FALSE, name = NULL, ...)
}
\description{
Forecaster Fit Arguments Helper function
}
