% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/spark-ml.R
\name{fitted.forecast_model}
\alias{fitted.forecast_model}
\alias{fitted.spark_ml}
\title{Forecast Model Fitted Method}
\usage{
\method{fitted}{forecast_model}(obj)

\method{fitted}{spark_ml}(obj)
}
\description{
Forecast Model Fitted Method

Fitted Method for Spark-ML Object
}
