% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{get_coefs.forecast_model}
\alias{get_coefs.forecast_model}
\title{Get Coefficients from Forecast Model Object}
\usage{
\method{get_coefs}{forecast_model}(mobj)
}
\arguments{
\item{mobj}{forecast model object as a result of fit function}
}
\description{
Get Coefficients from Forecast Model Object
}
