% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{get_forecasts.forecast_predict}
\alias{get_forecasts.forecast_predict}
\alias{get_forecasts.forecast}
\title{Function to Get Forecasts from forecast prediction}
\usage{
\method{get_forecasts}{forecast_predict}(pobj)

\method{get_forecasts}{forecast}(fobj)
}
\arguments{
\item{pobj}{forecast prediction object as a result of call to predict function}

\item{fobj}{forecast object}
}
\value{
data.frame with forecasts

data.frame with forecats from forecast model
}
\description{
Function to Get Forecasts from forecast prediction

Get Forecasts from forecast object
}
