% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_models}
\alias{get_models}
\title{Get All Modeler Models}
\usage{
get_models(obj, id = NULL, status = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{id}{vector of id strings}

\item{status}{model status code. accepts "created", "added", "trained", "evaluated", or "selected"}
}
\value{
list of model objects
}
\description{
Function to access all modeler models. Subsettable by either id or status
}
