% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{tidy_performance.forecast_model}
\alias{tidy_performance.forecast_model}
\title{Tidy Forecast Model Performance Object}
\usage{
\method{tidy_performance}{forecast_model}(mobj)
}
\arguments{
\item{mobj}{forecast model object as a result of fit function}
}
\description{
Tidy Forecast Model Performance Object
}
