% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{modeler}
\alias{modeler}
\title{Modeler Helper Function}
\usage{
modeler(df, target, type, name = NULL, id = NULL, version = NULL,
  desc = NULL, scientist = NULL, dir = NULL, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{target}{target variable for supervised use cases such as forecaster,
regresser, and classifier}

\item{type}{type of use case (forecaster, regresser, classifier, and
segmenter)}

\item{name}{optional name for use case. Default is to append target with type}

\item{id}{optional id string. A random id will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{dir}{optional directory path for saving reports, data and models}

\item{...}{any addtional parameters}
}
\value{
modeler object
}
\description{
Function to create a valid modeler object to support model building use
cases. Modeler is the base class from which the other more specific use cases
like forecaster, regresser, segmenter, etc... inherit form
}
\seealso{
Other use cases: \code{\link{new_forecaster}}
}
\concept{use cases}
