% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{modeler}
\alias{modeler}
\title{Modeler Helper Function}
\usage{
modeler(df, target, type, name = NULL, id = NULL, version = NULL,
  desc = NULL, scientist = NULL, ...)
}
\description{
Modeler Helper Function
}
