% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/forecaster.R
\name{predict.forecast_model}
\alias{predict.forecast_model}
\alias{predict.forecaster}
\title{Forecast Prediction Method}
\usage{
\method{predict}{forecast_model}(obj, periods, data = NULL, level = c(80,
  95))

\method{predict}{forecaster}(obj, periods, data = NULL, level = c(80, 95),
  desc = "")
}
\description{
Forecast Prediction Method

Method makes predictions for Forecaster's final model
}
