% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{predict.forecast_model}
\alias{predict.forecast_model}
\alias{predict}
\title{Forecast Prediction Method}
\usage{
\method{predict}{forecast_model}(obj, periods, data, level, ...)

predict(...)
}
\description{
Make model predictions
}
