% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{predict.forecast_model}
\alias{predict.forecast_model}
\alias{predict}
\title{Predict Forecast method}
\usage{
\method{predict}{forecast_model}(mobj, ...)

predict(...)
}
\arguments{
\item{mobj}{forecast model object as a result of fit function}
}
\description{
Method to make forecast predictions using a forecaster model object

Make model predictions
}
