% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{resample}
\alias{resample}
\alias{resample.numeric}
\alias{resample.data.frame}
\alias{resample.tbl_spark}
\title{Random Sample function}
\usage{
resample(x, number, amount, seed)

\method{resample}{numeric}(x, number, amount, seed = NULL)

\method{resample}{data.frame}(x, number, amount, seed = NULL)

\method{resample}{tbl_spark}(x, amount, seed = NULL)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create holdout samples from}

\item{number}{number of resamples to create}

\item{amount}{percent of data to randomly sample}

\item{seed}{optional input for setting random seed}
}
\description{
Function creates a random sample index from a dataset. Index can be used to
create data samples
}
