% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{rmse}
\alias{rmse}
\alias{rmse.data.frame}
\alias{rmse.tbl_spark}
\title{Generic rmse function}
\usage{
rmse(...)

\method{rmse}{data.frame}(x, predicted, actual)

\method{rmse}{tbl_spark}(x, predicted, actual)
}
\description{
Generic rmse function
}
