% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{roller_args}
\alias{roller_args}
\title{Roller Argument Helper Function}
\usage{
roller_args(order_vars, group_vars, measure_vars, fun, width, by, partial, mode,
  ...)
}
\description{
Creates a valid object of roller_args class
}
\details{
Function should be used in roller internals
}
