% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{summariser}
\alias{summariser}
\alias{summariser.data.frame}
\alias{summariser.tbl_spark}
\title{DataFrame Summariser function}
\usage{
summariser(df, ...)

\method{summariser}{data.frame}(df, group_vars, measure_vars = NULL,
  fun = c("sum"), ...)

\method{summariser}{tbl_spark}(df, group_vars, measure_vars = NULL,
  fun = c("sum"), ...)
}
\arguments{
\item{df}{DataFrame}

\item{...}{additional parameters passed to aggregate function}

\item{group_vars}{vector of column names. used in the group by aggregration.
can be zero or more columns. set to NULL for no grouping}

\item{measure_vars}{vector of column names. colnames to aggregrate. can be one
or more columns. Default is NULL which will calculate counts only}

\item{fun}{vector of function names. can be one or more columns. default is
sum. Current functions supported: n_distinct, min, max, sum, mean, variance,
sd, kurtosis, skewness, and percentile}
}
\value{
returns DataFrame with one row for each unique combination of grouping
 variables provided
}
\description{
Aggregation function. Configuarable inputs for grouping variables, measure
variables and aggregation functions.
}
\examples{
library(dplyr)


summariser(iris,
          group_vars = c("Species"),
          measure_vars = c("Sepal.Length", "Sepal.Width"),
          fun = c("sum", "mean"))
}
