% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{tidy_performance}
\alias{tidy_performance}
\title{Tidy Model Performance Generic}
\usage{
tidy_performance(obj)
}
\description{
Function to extract model performance on all samples. Converts output to
data.frame
}
