% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{train.forecast_model}
\alias{train.forecast_model}
\alias{train}
\title{Train Model Generic}
\usage{
\method{train}{forecast_model}(mobj, indicies, level)

train(...)
}
\value{
updated model object
}
\description{
Train single model to indicies provided
}
\details{
Fits Model and makes predictions for any validation or test indicies
provided. Adds fitted values and predictions to model's performance values
}
