% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R, R/modeler.R
\name{train_models.forecaster}
\alias{train_models.forecaster}
\alias{train_models}
\title{Train Model Generic}
\usage{
\method{train_models}{forecaster}(obj, ids = NULL)

train_models(obj, ...)
}
\description{
Train Model Generic
}
